refactor: resolved conflicts
