+++ packages/restapi/src/lib/space/Space.ts
@@ -33,7 +33,7 @@ import {
 } from '../types';
 import { VIDEO_CALL_TYPE } from '../payloads/constants';
 import sendLiveSpaceData from './helpers/sendLiveSpaceData';
-import { META_ACTION } from '../types/metaTypes';
+import { META_ACTION } from '../types/messageTypes';
 import { pCAIP10ToWallet } from '../helpers';
 
 export const initLiveSpaceData: LiveSpaceData = {
